Move scrapy spider out of .trac
